address feedback pt.4: extract embedding logic out of Embeddings service
