+++ lib/auth/init.go
@@ -39,7 +39,6 @@ import (
 	"github.com/gravitational/teleport/api/utils/keys"
 	apisshutils "github.com/gravitational/teleport/api/utils/sshutils"
 	"github.com/gravitational/teleport/lib"
-	"github.com/gravitational/teleport/lib/ai"
 	"github.com/gravitational/teleport/lib/auth/keystore"
 	"github.com/gravitational/teleport/lib/auth/native"
 	"github.com/gravitational/teleport/lib/backend"
@@ -233,10 +232,6 @@ type InitConfig struct {
 	// HTTPClientForAWSSTS overwrites the default HTTP client used for making
 	// STS requests. Used in test.
 	HTTPClientForAWSSTS utils.HTTPDoClient
-
-	// AIClient contains the client performing calls to AI services for Teleport Assist.
-	// In the Auth the first consumer is the embeddings service.
-	AIClient *ai.Client
 }
 
 // Init instantiates and configures an instance of AuthServer
