+++ lib/services/local/embeddings.go
@@ -16,7 +16,6 @@ package local
 
 import (
 	"context"
-	"crypto/sha256"
 	"time"
 
 	"github.com/gravitational/trace"
@@ -35,8 +34,7 @@ type EmbeddingsService struct {
 	log    *logrus.Entry
 	jitter retryutils.Jitter
 	backend.Backend
-	client ai.Embedder
-	clock  clockwork.Clock
+	clock clockwork.Clock
 }
 
 const (
@@ -84,71 +82,12 @@ func (e EmbeddingsService) UpsertEmbedding(ctx context.Context, embedding *ai.Em
 	return embedding, nil
 }
 
-// Embed takes a resource textual representation, checks if the resource
-// already has an up-to-date embedding stored in the backend, and computes
-// a new embedding otherwise. The newly computed embedding is stored in
-// the backend.
-func (e EmbeddingsService) Embed(ctx context.Context, kind string, resources map[string][]byte) ([]*ai.Embedding, error) {
-
-	// Lookup if there are embeddings in the backend for this node
-	// and the hash matches
-	embeddingsFromCache := make([]*ai.Embedding, 0)
-	toEmbed := make(map[string][]byte)
-	for name, data := range resources {
-		existingEmbedding, err := e.GetEmbedding(ctx, kind, name)
-		if err != nil && !trace.IsNotFound(err) {
-			return nil, trace.Wrap(err)
-		}
-		if err == nil {
-			if embeddingHashMatches(existingEmbedding, e.hash(data)) {
-				embeddingsFromCache = append(embeddingsFromCache, existingEmbedding)
-				continue
-			}
-		}
-		toEmbed[name] = data
-	}
-
-	// Convert to a list but keep track of the order so that we know which
-	// input maps to which resource.
-	keys := make([]string, 0, len(toEmbed))
-	input := make([]string, len(toEmbed))
-
-	for key := range toEmbed {
-		keys = append(keys, key)
-	}
-
-	for i, key := range keys {
-		input[i] = string(toEmbed[key])
-	}
-
-	response, err := e.client.ComputeEmbeddings(ctx, input)
-	if err != nil {
-		return nil, trace.Wrap(err)
-	}
-
-	newEmbeddings := make([]*ai.Embedding, 0, len(response))
-	for i, vector := range response {
-		newEmbeddings = append(newEmbeddings, ai.NewEmbedding(kind, keys[i], vector, e.hash(resources[keys[i]])))
-	}
-
-	// Store the new embeddings into the backend
-	for _, embedding := range newEmbeddings {
-		_, err := e.UpsertEmbedding(ctx, embedding)
-		if err != nil {
-			return nil, trace.Wrap(err)
-		}
-	}
-
-	return append(embeddingsFromCache, newEmbeddings...), nil
-}
-
 // NewEmbeddingsService is a constructor for the EmbeddingsService.
-func NewEmbeddingsService(b backend.Backend, embedder ai.Embedder) *EmbeddingsService {
+func NewEmbeddingsService(b backend.Backend) *EmbeddingsService {
 	return &EmbeddingsService{
 		log:     logrus.WithFields(logrus.Fields{trace.Component: "Embeddings"}),
 		jitter:  retryutils.NewFullJitter(),
 		Backend: b,
-		client:  embedder,
 		clock:   clockwork.NewRealClock(),
 	}
 }
@@ -156,16 +95,4 @@ func NewEmbeddingsService(b backend.Backend, embedder ai.Embedder) *EmbeddingsSe
 // embeddingItemKey builds the backend item key for a given ai.Embedding.
 func embeddingItemKey(embedding *ai.Embedding) []byte {
 	return backend.Key(embeddingsPrefix, embedding.GetName())
-}
-
-func embeddingHashMatches(embedding *ai.Embedding, hash ai.Sha256Hash) bool {
-	if len(embedding.EmbeddedHash) != 32 {
-		return false
-	}
-
-	return *(*ai.Sha256Hash)(embedding.EmbeddedHash) == hash
-}
-
-func (e EmbeddingsService) hash(data []byte) ai.Sha256Hash {
-	return sha256.Sum256(data)
 }
